requirements.txt
# Django e API
Django>=5.0,<6
djangorestframework>=3.15
django-environ>=0.11
django-cors-headers>=4.3
drf-spectacular>=0.27

# Banco
psycopg[binary]>=3.1

# HTTP / integrações
httpx>=0.27
cryptography>=42.0

# WhatsApp não oficial (cliente Baileys via servidor compatível)
pywabot>=0.3.9
